pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cappuccino"
version = "0.1.0"
description = "Asynchronous Cappuccino agent toolkit"
authors = [{name = "Cappuccino Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite",
    "aiohttp",
    "Pillow",
    "beautifulsoup4",
    "openai",
    "fastapi",
    "python-dotenv",
    "textblob",
    "httpx<0.28",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]
